Endrer navn på filer, funksjoner og variable slik at det stemmer bedre overens med hva det faktisk er
